Move local storage retry
