+++ agent/agent-tooling/src/test/java/com/azure/monitor/opentelemetry/exporter/implementation/localstorage/LocalFilePurgerTests.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.localstorage;
+package com.azure.monitor.opentelemetry.exporter.implementation.localstorage;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Collections.singletonList;
@@ -28,6 +28,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.io.File;
 import java.nio.ByteBuffer;
 import java.util.Collection;
+
+import com.azure.monitor.opentelemetry.exporter.implementation.localstorage.FileUtil;
+import com.azure.monitor.opentelemetry.exporter.implementation.localstorage.LocalFileCache;
+import com.azure.monitor.opentelemetry.exporter.implementation.localstorage.LocalFilePurger;
+import com.azure.monitor.opentelemetry.exporter.implementation.localstorage.LocalFileWriter;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
