+++ agent/agent-tooling/src/test/java/com/azure/monitor/opentelemetry/exporter/implementation/logging/NetworkFriendlyExceptionsTest.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.common;
+package com.azure.monitor.opentelemetry.exporter.implementation.logging;
 
 import static org.assertj.core.api.Assertions.assertThat;
