+++ agent/agent-tooling/src/test/java/com/microsoft/applicationinsights/agent/internal/common/TempDirsTest.java
@@ -26,7 +26,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.io.File;
 import org.junit.jupiter.api.Test;
 
-class LocalFileSystemUtilsTest {
+class TempDirsTest {
   /*
    * NOTE: it doesn't matter that Windows paths are converted to *nix paths and vice-versa.
    */
@@ -35,7 +35,7 @@ class LocalFileSystemUtilsTest {
   void getTempDir_WindowsForUserWithSystemUserUnknown() {
     final String input = "C:\\Users\\olivida\\AppData\\Local\\Temp";
 
-    File actual = LocalFileSystemUtils.getTempDir(input, "unknown");
+    File actual = TempDirs.getTempDir(input, "unknown");
 
     File expected = new File(input);
     assertThat(actual).isEqualTo(expected);
@@ -45,7 +45,7 @@ class LocalFileSystemUtilsTest {
   void getTempDir_WindowsForUserWithSystemUserDefined() {
     final String input = "C:\\Users\\olivida\\AppData\\Local\\Temp";
 
-    File actual = LocalFileSystemUtils.getTempDir(input, "olivida");
+    File actual = TempDirs.getTempDir(input, "olivida");
 
     File expected = new File(input);
     assertThat(actual).isEqualTo(expected);
@@ -55,7 +55,7 @@ class LocalFileSystemUtilsTest {
   void getTempDir_WindowsForUser() {
     final String input = "C:\\Users\\olivida\\AppData\\Local\\Temp";
 
-    File actual = LocalFileSystemUtils.getTempDir(input, null);
+    File actual = TempDirs.getTempDir(input, null);
 
     File expected = new File(input);
     assertThat(actual).isEqualTo(expected);
@@ -65,7 +65,7 @@ class LocalFileSystemUtilsTest {
   void getTempDir_Linux() {
     final String input = "/tmp";
 
-    File actual = LocalFileSystemUtils.getTempDir(input, "olivida");
+    File actual = TempDirs.getTempDir(input, "olivida");
 
     File expected = new File(input, "olivida").getAbsoluteFile();
     assertThat(actual).isEqualTo(expected);
@@ -75,7 +75,7 @@ class LocalFileSystemUtilsTest {
   void getTempDir_PerUser() {
     final String input = "/tmp/olivida";
 
-    File actual = LocalFileSystemUtils.getTempDir(input, "olivida");
+    File actual = TempDirs.getTempDir(input, "olivida");
 
     File expected = new File(input);
     assertThat(actual).isEqualTo(expected);
@@ -85,7 +85,7 @@ class LocalFileSystemUtilsTest {
   void getTempDir_InsideUserHome() {
     final String input = "/home/olivida/tmp";
 
-    File actual = LocalFileSystemUtils.getTempDir(input, "olivida");
+    File actual = TempDirs.getTempDir(input, "olivida");
 
     File expected = new File(input);
     assertThat(actual).isEqualTo(expected);
